+++ resources/css/workspace/views/list.css
@@ -1,273 +1,4 @@
-.list {
-  position: relative;
-}
-
-.list__styled-list,
-.list__styled-list ul {
-  list-style: none;
-  margin: 0;
-  padding: 0;
-}
-
-.list[data-showsearch='true'] .list__styled-list {
-  padding-top: 30px;
-}
-
-.list[data-folderopen='true'] .list__styled-list {
-  padding-top: 0;
-}
-
-.list--empty {
-  padding: 0.3rem 20px;
-}
-
-.list__styled-item {
-  display: flex;
-  margin: 0;
-  padding: 0;
-  touch-action: none;
-  position: relative;
-}
-
-.list__element {
-  color: var(--vscode-editorWidget-foreground);
-  display: flex;
-  flex-direction: column;
-  flex: 1;
-  font-size: var(--vscode-font-size);
-  margin: 0;
-  min-width: 0;
-  padding: 0.3rem 20px;
-  position: relative;
-  text-decoration: none;
-  touch-action: none;
-}
-
-.list__element--unselected {
-  cursor: pointer;
-}
-
-.list__element--selected {
-  background: var(--vscode-list-activeSelectionBackground);
-  color: var(--vscode-list-activeSelectionForeground);
-}
-
-@media (min-width: 300px) {
-  .list__element {
-    flex-direction: row;
-    flex-wrap: nowrap;
-    overflow: hidden;
-    text-overflow: ellipsis;
-    white-space: nowrap;
-  }
-}
-
-.list__element:hover {
-  background: var(--vscode-list-hoverBackground);
-  color: var(--vscode-editorWidget-foreground);
-}
-
-.list__element:focus {
-  outline: none;
-  background: var(--vscode-list-focusBackground);
-  color: var(--vscode-editorWidget-foreground);
-}
-
-.list__element:active {
-  color: var(--vscode-editorWidget-foreground);
-}
-
-.list__title {
-  display: block;
-  margin: 0;
-  overflow: hidden;
-  text-overflow: ellipsis;
-  white-space: nowrap;
-}
-
-.list__element--selected .list__title {
-  font-weight: bold;
-}
-
-@media (min-width: 300px) {
-  .list__title {
-    display: inline;
-    margin-right: 0.5rem;
-    overflow: unset;
-    text-overflow: unset;
-  }
-}
-
-.list__description {
-  color: var(--vscode-list-deemphasizedForeground);
-  display: block;
-  margin: 0;
-  margin-top: 0.25rem;
-  opacity: 0.75;
-  overflow: hidden;
-  text-overflow: ellipsis;
-  white-space: nowrap;
-}
-
-.list__element--selected > .list__description {
-  color: var(--vscode-list-activeSelectionForeground);
-  opacity: 1;
-}
-
-@media (min-width: 300px) {
-  .list__description {
-    display: inline;
-    margin-top: 0;
-  }
-}
-
-.list__element:hover > .list__description,
-.list__element:focus > .list__description,
-.list__element:active > .list__description {
-  opacity: 1;
-}
-
-.list__buttons {
-  align-items: center;
-  background: var(--vscode-list-hoverBackground);
-  display: none;
-  height: 100%;
-  justify-content: center;
-  line-height: 100%;
-  padding-left: 10px;
-  padding-right: 10px;
-  position: absolute;
-  right: 0;
-  top: 0;
-  white-space: nowrap;
-  width: 20px;
-}
-
-.list__element:hover > .list__buttons {
-  display: inline-flex;
-}
-
-.list__icon {
-  top: 0.25rem;
-}
-
-.list--empty .list__icon {
-  top: 0.3rem;
-}
-
-.list__search {
-  align-items: flex-start;
-  background: var(--vscode-sideBar-background);
-  box-sizing: border-box;
-  display: flex;
-  height: 30px;
-  justify-content: flex-start;
-  left: 0;
-  padding: 0 20px;
-  position: fixed;
-  top: 0;
-  width: 100%;
-  z-index: 1;
-}
-
-.list__search div {
-  width: 100%;
-}
-
-.list__search input {
-  background: var(--vscode-input-background);
-  border: 1px solid transparent;
-  color: var(--vscode-input-foreground);
-  font-size: var(--vscode-font-size);
-  padding: 4px;
-  width: 100%;
-}
-
-.list__search input:focus {
-  outline-color: var(--vscode-focusBorder);
-}
-
-.list__styled-list-searchedout {
-  color: var(--vscode-foreground);
-  padding: 0.3rem 20px;
-  padding-top: 30px;
-}
-
-.list__styled-list-searchedout p {
-  margin: 0;
-  padding: 0.3rem 0;
-}
-
-.list__folder {
-  position: relative;
-  background: rgba(0, 0, 0, 0.1);
-}
-
-.list[data-showsearch='true'] .list__folder {
-  padding-top: 30px;
-}
-
-.list__folder-title {
-  margin: 0.7rem 0;
-  padding: 0 20px;
-}
-
-.list__folder-controls {
-  padding-bottom: 0.7rem;
-  padding-top: 0.3rem;
-  text-align: center;
-}
-
-.list__branch-list-item .list__item::before {
-  background: rgba(169, 169, 169, 0.4);
-  content: '';
-  display: inline-block;
-  height: 100%;
-  left: 10px;
-  position: absolute;
-  top: 0;
-  width: 1px;
-  z-index: 2;
-}
-
-.list__branch-list-item .list__item:last-child::before {
-  height: 50%;
-}
-
-.list__branch-list-item > .list__element {
-  padding-bottom: 0;
-}
-
-.list__branch-list-item > .list__element > .list__title {
-  padding-bottom: 0.3rem;
-}
-
-.list__branch-list-item > .list__element .list__element {
-  padding-right: 0;
-}
-
-.list__folder .list__element:hover {
-  background: unset;
-}
-
-.list__folder .list__element:hover > .list__description,
-.list__folder .list__element:focus > .list__description,
-.list__folder .list__element:active > .list__description {
-  opacity: 0.75;
-}
-
-.list__styled-list--tree {
-  position: relative;
-}
-
-.list__branch-list-item {
-  position: relative;
-}
-
-.list__branch-list-item > .list__element:hover {
-  background: unset;
-}
-
-.list__branch-list-item > .list__element:focus {
-  background: unset;
-}
+@import '../snippets/list.css';
+@import ' ../snippets/search.css';
+@import '../snippets/folder.css';
+@import ' ../snippets/tree.css';
