+++ src/templates/workspace/snippets/list.ts
@@ -2,33 +2,8 @@ import { t } from 'vscode-ext-localisation';
 import { listItem } from '..';
 import { WorkspaceState } from '../../../webviews';
 import { RenderVars } from '../../../webviews/webviews.interface';
-import { FileTree, getFileTree } from '../../../webviews/Workspace/helpers/getFileTree';
-import { File } from '../../../webviews/Workspace/WorkspaceViewProvider.interface';
-
-export const isFile = (branch: FileTree | File): branch is File => {
-  return (branch as File).isSelected !== undefined;
-};
-
-const renderTree = (branch: FileTree, renderVars: RenderVars): string => {
-  return Object.entries(branch)
-    .map(([key, value]) => {
-      if (isFile(value)) {
-        return listItem(value, renderVars);
-      }
-
-      return `
-        <li class="list__branch-list-item list__styled-item">
-          <span class="list__element" title="${key}">
-            <span class="list__title">${key}</span>
-            <ul>
-              ${renderTree(value, renderVars)}
-            </ul>
-          </span>
-        </li>
-      `;
-    })
-    .join('');
-};
+import { getFileTree } from '../../../webviews/Workspace/helpers/getFileTree';
+import { tree } from './tree';
 
 export const list = (state: WorkspaceState, renderVars: RenderVars) => {
   const { convertedFiles, files, search, visibleFiles } = state;
@@ -55,7 +30,7 @@ export const list = (state: WorkspaceState, renderVars: RenderVars) => {
       <ul class="list__list list__styled-list${showTree ? ' list__styled-list--tree' : ''}">
         ${
           showTree
-            ? renderTree(getFileTree(convertedFiles), renderVars)
+            ? tree(getFileTree(convertedFiles), renderVars)
             : visibleFiles.map((file) => listItem(file, renderVars)).join('')
         }
       </ul>
