+++ src/templates/workspace/snippets/treeArrow.ts
@@ -1,0 +1,5 @@
+export const treeArrow = () => {
+  return `
+      <div class="list_branch-arrow codicon codicon-chevron-down"></div>
+    `;
+};
